requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.6
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
dill==0.3.5.1
Flask==2.1.2
frozenlist==1.3.0
idna==2.8
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.2
parso==0.8.3
platformdirs==2.5.2
PyJWT==2.4.0
pylint==2.14.3
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.22.0
tomli==2.0.1
tomlkit==0.11.0
twilio==7.10.0
typing_extensions==4.2.0
urllib3==1.25.11
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
